fullpotential_ai/fullpotential_core/droplets/hteam/droplet-1/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyJWT==2.9.0
psutil==6.0.0
prometheus-client==0.20.0
jsonschema==4.23.0
requests==2.32.3
pyyaml==6.0.2
python-dotenv==1.0.0
